@DATAJPATest

+++ src/main/java/com/Ankita/TestingApp/TestingAppApplication/entites/Employee.java
@@ -24,5 +24,5 @@ public class Employee {
 
     private String name;
 
-    private Double salary;
+    private Long salary;
 }
